# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AuthEngine",
    "description": "Production-ready IAM system built with FastAPI — supports authentication, multi-tenancy, fine-grained permissions, and token introspection to power multiple apps from a unified identity layer.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tenants/{tenant_id}/sms-config": {
      "get": {
        "tags": ["tenant-sms-config"],
        "summary": "Get tenant SMS configuration",
        "operationId": "get_sms_config_api_v1_tenants__tenant_id__sms_config_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/TenantSMSConfigResponse" },
                    {
                      "$ref": "#/components/schemas/TenantSMSConfigFallbackResponse"
                    }
                  ],
                  "title": "Response Get Sms Config Api V1 Tenants  Tenant Id  Sms Config Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["tenant-sms-config"],
        "summary": "Create tenant SMS configuration",
        "operationId": "create_sms_config_api_v1_tenants__tenant_id__sms_config_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TenantSMSConfigCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantSMSConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["tenant-sms-config"],
        "summary": "Update tenant SMS configuration",
        "operationId": "update_sms_config_api_v1_tenants__tenant_id__sms_config_put",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TenantSMSConfigUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantSMSConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["tenant-sms-config"],
        "summary": "Delete tenant SMS configuration (falls back to platform default)",
        "operationId": "delete_sms_config_api_v1_tenants__tenant_id__sms_config_delete",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/sms-config/test": {
      "post": {
        "tags": ["tenant-sms-config"],
        "summary": "Send a test SMS using the tenant's configured provider",
        "operationId": "test_sms_config_api_v1_tenants__tenant_id__sms_config_test_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SMSConfigTestRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SMSConfigTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SMSConfigTestRequest": {
        "properties": {
          "to_number": {
            "type": "string",
            "minLength": 1,
            "title": "To Number"
          }
        },
        "type": "object",
        "required": ["to_number"],
        "title": "SMSConfigTestRequest",
        "description": "Body for POST /tenants/{tenant_id}/sms-config/test"
      },
      "SMSConfigTestResponse": {
        "properties": {
          "success": { "type": "boolean", "title": "Success" },
          "error": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Error"
          }
        },
        "type": "object",
        "required": ["success"],
        "title": "SMSConfigTestResponse"
      },
      "SMSProviderType": {
        "type": "string",
        "enum": ["twilio", "console"],
        "title": "SMSProviderType"
      },
...
      "TenantSMSConfigCreate": {
        "properties": {
          "provider": { "$ref": "#/components/schemas/SMSProviderType" },
          "api_key": {
            "type": "string",
            "minLength": 1,
            "title": "Api Key",
            "description": "Raw API key / auth token — encrypted"
          },
          "from_number": {
            "type": "string",
            "minLength": 1,
            "title": "From Number"
          },
          "account_sid": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Account Sid",
            "description": "Twilio account SID (if provider is twilio)"
          }
        },
        "type": "object",
        "required": ["provider", "api_key", "from_number"],
        "title": "TenantSMSConfigCreate",
        "description": "Body for POST /tenants/{tenant_id}/sms-config"
      },
      "TenantSMSConfigFallbackResponse": {
        "properties": {
          "configured": {
            "type": "boolean",
            "title": "Configured",
            "default": false
          },
          "using_platform_default": {
            "type": "boolean",
            "title": "Using Platform Default",
            "default": true
          },
          "platform_provider": {
            "type": "string",
            "title": "Platform Provider"
          },
          "platform_from_number": {
            "type": "string",
            "title": "Platform From Number"
          }
        },
        "type": "object",
        "required": ["platform_provider", "platform_from_number"],
        "title": "TenantSMSConfigFallbackResponse",
        "description": "Returned when tenant has no custom config."
      },
      "TenantSMSConfigResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "provider": { "type": "string", "title": "Provider" },
          "from_number": { "type": "string", "title": "From Number" },
          "credential_hint": { "type": "string", "title": "Credential Hint" },
          "account_sid": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Account Sid"
          },
          "is_active": { "type": "boolean", "title": "Is Active" }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "provider",
          "from_number",
          "credential_hint",
          "is_active"
        ],
        "title": "TenantSMSConfigResponse",
        "description": "Response — never includes the real API key."
      },
      "TenantSMSConfigUpdate": {
        "properties": {
          "provider": {
            "anyOf": [
              { "$ref": "#/components/schemas/SMSProviderType" },
              { "type": "null" }
            ]
          },
          "api_key": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Api Key",
            "description": "If omitted, existing key is kept"
          },
          "from_number": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "From Number"
          },
          "account_sid": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Account Sid"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TenantSMSConfigUpdate",
        "description": "Body for PUT /tenants/{tenant_id}/sms-config"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}